Don't overwrite existing catalog json

+++ get_library_catalog.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import os
+from os.path import exists
 import sys
 appdir=os.path.dirname(os.path.abspath(__file__))
 os.chdir(appdir)
@@ -29,14 +30,16 @@ if __name__ == "__main__":
 
     assert_env_vars()
     target_dir = os.getenv('target_dir')
+    os.makedirs(f'{target_dir}/catalog', 0o0700, exist_ok=True)
 
     for account in get_auth_creds():
         cc = account.locale.country_code
         ab = account.activation_bytes
-        os.makedirs(f'{target_dir}/catalog', 0o0700, exist_ok=True)
-        os.makedirs(f'{target_dir}/catalog/{cc}.{ab}', 0o0700, exist_ok=True)
 
         with audible.Client(auth=account) as client:
+            catalog_dir = f'{target_dir}/catalog/{cc}.{ab}'
+            os.makedirs(catalog_dir, 0o0700, exist_ok=True)
+
             library = client.get(
                 "1.0/library",
                 num_results=1,
@@ -46,6 +49,12 @@ if __name__ == "__main__":
             for book in library["items"]:
                 asin = book['asin']
                 print(f"found '{book['title']}' by {book['authors'][0]['name']}")
-                catalog_file = f"{target_dir}/catalog/{cc}.{ab}/{asin}.json"
+                book['_STATE'] = {}
+
+                catalog_file = f"{catalog_dir}/{asin}.json"
+                if exists(catalog_file):
+                    print(f"** SKIPPING {asin}, catalog json already exists")
+                    continue
+
                 with open(catalog_file, 'w') as f:
                     f.write(json.dumps(book))
